Fix conflict

+++ Sources/Framework/Common/Memory/Allocator.h
@@ -73,9 +73,9 @@ namespace ProjectEngine
         size_t      m_nBlocksPerPage;
 
         // statistics
-        size_t    m_nPages;
-        size_t    m_nBlocks;
-        size_t    m_nFreeBlocks;
+        uint32_t    m_nPages;
+        uint32_t    m_nBlocks;
+        uint32_t    m_nFreeBlocks;
 
         // disable copy & assignment
         Allocator(const Allocator& clone);
